Reworked text censoring functions
Single function now splits work into 2 subfunctions. The 1st function checks if a word is the same as the offending text with spaces before or after it. The 2nd function tries to censor every appearance of the offending text that does not have an alphabet character before and after it.